+++ addOns/graphql/CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this add-on will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.1.0] - 2020-08-28
+- First Version
+- Features
+  - Import a GraphQL Schema
+  - Generate Queries from an imported Schema
 
-- Bare-bones add-on (no functionality yet).
+[0.1.0]: https://github.com/zaproxy/zap-extensions/releases/graphql-v0.1.0